requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.7
django-debug-toolbar==4.2.0
django-embed-video==1.4.9
django-recaptcha==4.0.0
idna==3.4
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==10.1.0
protobuf==4.21.12
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
setuptools==70.0.0